pyproject.toml
[project]
name = "cv-search-poc"
version = "0.1.0"
description = "Candidate search and planning tools"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.7",
    "python-dotenv>=1.0.1",
    "pydantic>=2.8.2",
    "pydantic-settings>=2.0.0",
    "openai>=1.42.0,<2",
    "python-pptx",
    "redis>=5.0.0",
    "watchdog>=4.0.0",
    "psycopg[binary,pool]>=3.2",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
]

[project.optional-dependencies]
streamlit = [
    "streamlit[auth]>=1.37.0",
    "Authlib>=1.3.2,<1.4",
]
dev = [
    "ruff>=0.6.0",
    "pytest>=7.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cv_search"]

[tool.hatch.build.targets.sdist]
include = [
    "src/cv_search",
    "README.md",
    "pyproject.toml",
    "app.py",
    "main.py",
    "pages/**",
    "data/**",
    ".env.example",
]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F"]
ignore = [
    "E203",
    "E501",
]
